# packages/contracts/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Lingua-Bloom Source Ingestion API
  version: 0.3.0
paths:
...
  /api/imports/{runId}:
    get:
      summary: Return the owner's run state, ordered events, issues, and draft reference
      security:
        - teacherSession: []
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Current import state
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ImportStatus' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { description: Run not found }
...
  /api/imports/{runId}/review:
    post:
      summary: Submit idempotent teacher decisions and resume processing
      security:
        - teacherSession: []
      parameters:
        - $ref: '#/components/parameters/RunId'
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/ReviewSubmission' }
      responses:
        '202': { description: Review accepted }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
        '409':
          description: Idempotency conflict or stale draft version
          content:
            application/json:
              schema:
                oneOf:
                  - $ref: '#/components/schemas/DraftVersionConflict'
                  - $ref: '#/components/schemas/IdempotencyConflict'
...
  /api/lessons/{lessonId}/versions:
    get:
      summary: List immutable lesson versions available to the owner
      security:
        - teacherSession: []
      parameters:
        - $ref: '#/components/parameters/LessonId'
      responses:
        '200':
          description: Ordered lesson versions
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/PublishedVersion' }
        '401': { $ref: '#/components/responses/Unauthenticated' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    teacherSession:
      type: apiKey
      in: cookie
      name: sb-auth-token
      description: Supabase SSR project-scoped auth cookie; the deployed cookie may be chunked
  parameters:
    RunId:
      in: path
      name: runId
      required: true
      schema: { type: string, minLength: 1 }
    LessonId:
      in: path
      name: lessonId
      required: true
      schema: { type: string, minLength: 1 }
...
  responses:
...
    Unauthenticated:
      description: Valid teacher session required
    Forbidden:
      description: Authenticated teacher does not own the requested resource
...
  schemas:
...
    ImportStatus:
      type: object
      additionalProperties: false
      required: [runId, status, updatedAt, recovery, events, issues, failure]
      properties:
        runId: { type: string }
        status:
          enum: [accepted, processing, awaiting_review, blocked, ready_to_publish, completed, cancelled, failed]
        updatedAt: { type: string, format: date-time }
        recovery:
          oneOf:
            - type: "null"
            - $ref: "#/components/schemas/StaleRunRecovery"
        events: { type: array, items: { type: object } }
        issues: { type: array, items: { type: object } }
        draftId: { type: [string, 'null'] }
        failure:
          oneOf:
            - type: 'null'
            - $ref: '#/components/schemas/FailureInfo'
      allOf:
        - if:
            properties: { status: { const: failed } }
          then:
            properties:
              failure: { $ref: '#/components/schemas/FailureInfo' }
        - if:
            properties: { status: { const: blocked } }
          then:
            properties: { issues: { minItems: 1 } }
        - if:
            properties:
              status:
                enum: [accepted, processing, awaiting_review, blocked, ready_to_publish, completed, cancelled]
          then:
            properties: { failure: { type: 'null' } }
    FailureInfo:
      type: object
      additionalProperties: false
      required: [code, kind, message, manualResumeAllowed]
      properties:
        code: { type: string, minLength: 1 }
        kind: { enum: [retriable, terminal] }
        message: { type: string, minLength: 1 }
        manualResumeAllowed: { type: boolean }
        limitType: { enum: [pdfPages, pdfBytes, textCharacters, answerFields] }
        limit: { type: integer, minimum: 1 }
        actual: { type: integer, minimum: 1 }
      allOf:
        - if:
            properties: { kind: { const: retriable } }
          then:
            properties: { manualResumeAllowed: { const: true } }
        - if:
            properties: { kind: { const: terminal } }
          then:
            properties: { manualResumeAllowed: { const: false } }
        - if:
            properties: { code: { const: SOURCE_TOO_LARGE } }
          then:
            required: [limitType, limit, actual]
...
    DraftVersionConflict:
      type: object
      additionalProperties: false
      required: [code, message, currentDraftVersion]
      properties:
        code: { const: DRAFT_VERSION_CONFLICT }
        message: { type: string, minLength: 1 }
        currentDraftVersion: { type: integer, minimum: 1 }
    IdempotencyConflict:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { const: IDEMPOTENCY_CONFLICT }
        message: { type: string, minLength: 1 }
    StaleRunRecovery:
      type: object
      additionalProperties: false
      required: [kind, redispatchAllowed, staleSince]
      properties:
        kind: { enum: [dispatch_not_started, worker_heartbeat_expired] }
        redispatchAllowed: { const: true }
        staleSince: { type: string, format: date-time }
...
    ReviewSubmission:
      type: object
      additionalProperties: false
      required: [draftVersion, idempotencyKey]
      anyOf:
        - properties: { decisions: { minItems: 1 } }
        - properties: { answerReviews: { minItems: 1 } }
        - properties: { exerciseEdits: { minItems: 1 } }
      properties:
        draftVersion: { type: integer, minimum: 1 }
        idempotencyKey: { type: string, minLength: 16, maxLength: 128 }
        decisions:
          type: array
          default: []
          items:
            type: object
            additionalProperties: false
            required: [issueId, decision, reason]
            properties:
              issueId: { type: string, format: uuid }
              decision: { enum: [confirm, edit, exclude] }
              reason: { type: string, minLength: 1 }
              replacementValue: { type: [string, "null"] }
        answerReviews:
          type: array
          default: []
          items:
            type: object
            additionalProperties: false
            required: [answerFieldId, decision, reason, replacementValue]
            properties:
              answerFieldId: { type: string, minLength: 1 }
              issueId: { type: [string, "null"], format: uuid }
              decision: { enum: [confirm, edit] }
              reason: { type: string, minLength: 1 }
              replacementValue: { type: string, minLength: 1 }
        exerciseEdits:
          type: array
          default: []
          items:
            type: object
            additionalProperties: false
            required: [exerciseId, prompt, options]
            properties:
              exerciseId: { type: string, minLength: 1 }
              prompt: { type: string, minLength: 1 }
              options:
                type: array
                items:
                  type: object
                  additionalProperties: false
                  required: [id, value]
                  properties:
                    id: { type: string, minLength: 1 }
                    value: { type: string }
...
    PublishedVersion:
      type: object
      additionalProperties: false
      required: [lessonId, publicLessonId, version]
      properties:
        lessonId: { type: string }
        publicLessonId:
          type: string
          minLength: 22
          pattern: '^[A-Za-z0-9_-]+$'
        version: { type: integer, minimum: 1 }